Using OOP to save duplicate code
