+++ src/Omniphx/Forrest/Providers/Lumen/ForrestServiceProvider.php
@@ -2,80 +2,23 @@
 
 namespace Omniphx\Forrest\Providers\Lumen;
 
-use GuzzleHttp\Client;
-use Illuminate\Support\ServiceProvider;
-use Omniphx\Forrest\Providers\Laravel\LaravelCache;
-use Omniphx\Forrest\Providers\Laravel\LaravelEvent;
-use Omniphx\Forrest\Providers\Laravel\LaravelInput;
-use Omniphx\Forrest\Providers\Laravel\LaravelRedirect;
-use Omniphx\Forrest\Providers\Laravel\LaravelSession;
+use Omniphx\Forrest\Providers\BaseServiceProvider;
 
-class ForrestServiceProvider extends ServiceProvider
+class ForrestServiceProvider extends BaseServiceProvider
 {
     /**
-     * Indicates if loading of the provider is deferred.
+     * Indicates if the application is laravel/lumen
      *
      * @var bool
      */
-    protected $defer = false;
+    protected $is_laravel = false;
 
     /**
-     * Bootstrap the application events.
+     * Returns the location of the package config file
      *
-     * @return void
+     * @return String file location
      */
-    public function boot()
-    {
-        $this->publishes([
-            __DIR__.'/../../../../config/config.php' => $this->configPath(),
-        ]);
-    }
-
-    /**
-     * Register the service provider.
-     *
-     * @return void
-     */
-    public function register()
-    {
-        $this->app->singleton('forrest', function ($app) {
-
-            //Config options:
-            $settings = config('forrest');
-            $storageType = config('forrest.storage.type');
-            $authenticationType = config('forrest.authentication');
-
-            //Dependencies:
-            $client = new Client();
-            $input = new LaravelInput();
-            $event = new LaravelEvent();
-            $redirect = new LaravelRedirect();
-
-            //Determine storage dependency:
-            if ($storageType == 'cache') {
-                $storage = new LaravelCache(app('config'), app('cache'));
-            } else {
-                $storage = new LaravelSession(app('config'), app('session'));
-            }
-
-            //Class namespace:
-            $forrest = "\\Omniphx\\Forrest\\Authentications\\$authenticationType";
-
-            return new $forrest($client, $event, $input, $redirect, $storage, $settings);
-        });
-    }
-
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
-    {
-        return [];
-    }
-
-    protected function configPath()
+    protected function getConfigPath()
     {
         return __DIR__.'/../config/forrest.php';
     }
